USB driver cleanup
