+++ radio/src/targets/common/arm/stm32/CMakeLists.txt
@@ -36,7 +36,7 @@ set(FIRMWARE_TARGET_SRC
   ../common/arm/stm32/usb_bsp.c
   ../common/arm/stm32/usbd_desc.c
   ../common/arm/stm32/usbd_usr.cpp
-  ../common/arm/stm32/usb_driver.c
+  ../common/arm/stm32/usb_driver.cpp
   )
 if(USB STREQUAL SERIAL)
   add_definitions(-DUSB_SERIAL)
