+++ radio/src/targets/common/arm/stm32/usb_driver.cpp
@@ -18,9 +18,16 @@
  * GNU General Public License for more details.
  */
 
-#include "board.h"
-#include "usb_dcd_int.h"
-#include "usb_bsp.h"
+#if defined(__cplusplus)
+  extern "C" {
+#endif
+  #include "usb_dcd_int.h"
+  #include "usb_bsp.h"
+#if defined(__cplusplus)
+  }
+#endif
+
+#include "opentx.h"
 #include "debug.h"
 
 int usbPlugged(void)
@@ -46,7 +53,7 @@ int usbPlugged(void)
 
 USB_OTG_CORE_HANDLE USB_OTG_dev;
 
-void OTG_FS_IRQHandler(void)
+extern "C" void OTG_FS_IRQHandler(void)
 {
   DEBUG_INTERRUPT(INT_OTG_FS);
   USBD_OTG_ISR_Handler(&USB_OTG_dev);
@@ -76,3 +83,46 @@ void usbStop(void)
 {
   USBD_DeInit(&USB_OTG_dev);
 }
+
+#if defined(USB_JOYSTICK)
+
+/*
+  Prepare and send new USB data packet
+
+  The format of HID_Buffer is defined by
+  USB endpoint description can be found in
+  file usb_hid_joystick.c, variable HID_JOYSTICK_ReportDesc
+*/
+void usbJoystickUpdate(void)
+{
+  static uint8_t HID_Buffer[HID_IN_PACKET];
+
+  //buttons
+  HID_Buffer[0] = 0;
+  HID_Buffer[1] = 0;
+  HID_Buffer[2] = 0;
+  for (int i = 0; i < 8; ++i) {
+    if ( channelOutputs[i+8] > 0 ) {
+      HID_Buffer[0] |= (1 << i);
+    }
+    if ( channelOutputs[i+16] > 0 ) {
+      HID_Buffer[1] |= (1 << i);
+    }
+    if ( channelOutputs[i+24] > 0 ) {
+      HID_Buffer[2] |= (1 << i);
+    }
+  }
+
+  //analog values
+  //uint8_t * p = HID_Buffer + 1;
+  for (int i = 0; i < 8; ++i) {
+    int16_t value = channelOutputs[i] / 8;
+    if ( value > 127 ) value = 127;
+    else if ( value < -127 ) value = -127;
+    HID_Buffer[i+3] = static_cast<int8_t>(value);
+  }
+
+  USBD_HID_SendReport (&USB_OTG_dev, HID_Buffer, HID_IN_PACKET );
+}
+
+#endif // #defined(USB_JOYSTICK)
