+++ radio/src/targets/taranis/board.cpp
@@ -264,51 +264,6 @@ void boardOff()
   pwrOff();
 }
 
-#if defined(USB_JOYSTICK) && !defined(SIMU)
-extern USB_OTG_CORE_HANDLE USB_OTG_dev;
-
-/*
-  Prepare and send new USB data packet
-
-  The format of HID_Buffer is defined by
-  USB endpoint description can be found in
-  file usb_hid_joystick.c, variable HID_JOYSTICK_ReportDesc
-*/
-void usbJoystickUpdate(void)
-{
-  static uint8_t HID_Buffer[HID_IN_PACKET];
-
-  //buttons
-  HID_Buffer[0] = 0;
-  HID_Buffer[1] = 0;
-  HID_Buffer[2] = 0;
-  for (int i = 0; i < 8; ++i) {
-    if ( channelOutputs[i+8] > 0 ) {
-      HID_Buffer[0] |= (1 << i);
-    }
-    if ( channelOutputs[i+16] > 0 ) {
-      HID_Buffer[1] |= (1 << i);
-    }
-    if ( channelOutputs[i+24] > 0 ) {
-      HID_Buffer[2] |= (1 << i);
-    }
-  }
-
-  //analog values
-  //uint8_t * p = HID_Buffer + 1;
-  for (int i = 0; i < 8; ++i) {
-    int16_t value = channelOutputs[i] / 8;
-    if ( value > 127 ) value = 127;
-    else if ( value < -127 ) value = -127;
-    HID_Buffer[i+3] = static_cast<int8_t>(value);
-  }
-
-  USBD_HID_SendReport (&USB_OTG_dev, HID_Buffer, HID_IN_PACKET );
-}
-
-#endif // #defined(USB_JOYSTICK) && !defined(SIMU)
-
-
 uint8_t currentTrainerMode = 0xff;
 
 void checkTrainerSettings()
